+++ src/Avalonia.Samples/Drawing/BattleCity/README.md
@@ -7,10 +7,12 @@ Port of https://github.com/hacklex/PekaCity to Avalonia.
 
 ## What is this? Why?
 
-Well, this is a stub for a 2D game. The purpose of the project was to demonstrate that one can write a 2D game in AvaloniaUI without writing any rendering code.
+This is a stub for a 2D game. The purpose of the project was to demonstrate that one can write a 2D game in AvaloniaUI without writing any rendering code.
+
 ## Features
 
-- 2D Tiles. Not yet animated, but animating won't be a problem, I guess
+- 2D Tiles
 - Cell-aligned game objects
+- Has basic implementation of a typical game loop
 - No rendering code, everything is done using AvaloniaUI data binding and a few ValueConverters
